+++ testapi/tests/test_views.py
@@ -2,9 +2,13 @@ import pytest
 from rest_framework.test import APIClient
 
 from django.urls import reverse
+from factory import Sequence
+from rest_framework import status
 
 from client.tests.factories import ClientFactory
-from submission import models
+from submission import constants, models
+from submission.models import Submission
+from submission.tests.factories import SubmissionFactory, SenderFactory
 
 
 @pytest.fixture
@@ -14,8 +18,9 @@ def user():
 
 def api_client(settings, user, test_api_flag):
     settings.SIGAUTH_URL_NAMES_WHITELIST = [
-        'submissions-by-email',
-        'submission'
+        'delete_test_submissions',
+        'submissions_by_email',
+        'submission',
     ]
     settings.FEATURE_TEST_API_ENABLED = test_api_flag
     client = APIClient()
@@ -55,15 +60,14 @@ def test_find_submissions_by_email(api_client_enabled_test_api):
         format='json'
     )
 
-    assert response.status_code == 201
+    assert response.status_code == status.HTTP_201_CREATED
     assert models.Submission.objects.count() == 1
 
     response = api_client_enabled_test_api.get(
-        reverse('testapi:submissions-by-email',
-                kwargs={'email_address': 'foo@bar.com'}),
+        reverse('testapi:submissions_by_email', kwargs={'email_address': 'foo@bar.com'}),
         format='json'
     )
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     assert response.json()[0]['meta']['recipients'][0] == 'foo@bar.com'
     assert response.json()[0]['is_sent'] is True
 
@@ -73,11 +77,11 @@ def test_return_404_if_no_submissions_are_found(api_client_enabled_test_api):
     assert models.Submission.objects.count() == 0
 
     response = api_client_enabled_test_api.get(
-        reverse('testapi:submissions-by-email',
+        reverse('testapi:submissions_by_email',
                 kwargs={'email_address': 'foo@bar.com'}),
         format='json'
     )
-    assert response.status_code == 404
+    assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
 @pytest.mark.django_db
@@ -85,17 +89,113 @@ def test_return_404_if_testapi_is_disabled(api_client_disabled_testapi):
     assert models.Submission.objects.count() == 0
 
     response = api_client_disabled_testapi.get(
-        reverse('testapi:submissions-by-email',
+        reverse('testapi:submissions_by_email',
                 kwargs={'email_address': 'foo@bar.com'}),
         format='json'
     )
-    assert response.status_code == 404
+    assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
 def test_return_401_when_unauthenticated():
     client = APIClient()
     response = client.get(
-        reverse('testapi:submissions-by-email',
+        reverse('testapi:submissions_by_email',
                 kwargs={'email_address': 'foo@bar.com'})
     )
-    assert response.status_code == 401
+    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+
+
+@pytest.mark.django_db
+def test_delete_test_submissions(api_client_enabled_test_api):
+    SubmissionFactory.create_batch(
+        3,
+        sender=SenderFactory(
+            email_address=Sequence(lambda n: f'test+{n}@directory.uktrade.io')
+        )
+    )
+    response = api_client_enabled_test_api.delete(
+        reverse('testapi:delete_test_submissions')
+    )
+    assert response.status_code == status.HTTP_204_NO_CONTENT
+    assert Submission.objects.count() == 0
+
+
+@pytest.mark.django_db
+def test_delete_test_submissions_email_action(api_client_enabled_test_api):
+    meta = {
+        'action_name': constants.ACTION_NAME_EMAIL,
+        'recipients': ['test+notification@directory.uktrade.io'],
+        'form_url': '/the/form/tests',
+        'funnel_steps': ['one', 'two', 'three'],
+        'reply_to': 'test@testsubmission.com',
+    }
+    SubmissionFactory.create_batch(3, meta=meta)
+
+    response = api_client_enabled_test_api.delete(
+        reverse('testapi:delete_test_submissions')
+    )
+    assert response.status_code == status.HTTP_204_NO_CONTENT
+    assert Submission.objects.count() == 0
+
+
+@pytest.mark.django_db
+def test_delete_test_submissions_gov_notify_action(
+        api_client_enabled_test_api, gov_notify_email_action_payload
+):
+    email = 'test+govnotifyaction@directory.uktrade.io'
+    gov_notify_email_action_payload['meta']['sender']['email_address'] = email
+    Submission.objects.create(**gov_notify_email_action_payload)
+
+    response = api_client_enabled_test_api.delete(
+        reverse('testapi:delete_test_submissions')
+    )
+    assert response.status_code == status.HTTP_204_NO_CONTENT
+    assert Submission.objects.count() == 0
+
+
+@pytest.mark.django_db
+def test_delete_test_submissions_zendesk_action(
+        api_client_enabled_test_api, zendesk_action_payload
+):
+    email = 'test+zendeskaction@directory.uktrade.io'
+    zendesk_action_payload['meta']['email_address'] = email
+    Submission.objects.create(**zendesk_action_payload)
+
+    response = api_client_enabled_test_api.delete(
+        reverse('testapi:delete_test_submissions')
+    )
+    assert response.status_code == status.HTTP_204_NO_CONTENT
+    assert Submission.objects.count() == 0
+
+
+@pytest.mark.django_db
+def test_delete_test_submissions_returns_404_when_no_test_submissions(
+        api_client_enabled_test_api
+):
+    response = api_client_enabled_test_api.delete(
+        reverse('testapi:delete_test_submissions')
+    )
+    assert response.status_code == status.HTTP_404_NOT_FOUND
+
+
+@pytest.mark.django_db
+def test_delete_test_submissions_returns_404_with_disabled_testapi(
+        api_client_disabled_testapi
+):
+    SubmissionFactory.create(
+        sender=SenderFactory(
+            email_address=Sequence(lambda n: f'test+{n}@directory.uktrade.io')
+        )
+    )
+    response = api_client_disabled_testapi.delete(
+        reverse('testapi:delete_test_submissions')
+    )
+    assert response.status_code == status.HTTP_404_NOT_FOUND
+    assert Submission.objects.count() == 1
+
+
+@pytest.mark.django_db
+def test_delete_test_submissions_returns_401_when_unauthenticated():
+    client = APIClient()
+    response = client.delete(reverse('testapi:delete_test_submissions'))
+    assert response.status_code == status.HTTP_401_UNAUTHORIZED